Rename formatter to more relevant names
