+++ src/format-message-friendly.ts
