+++ test/parsing-utils-friendly.test.ts
@@ -16,7 +16,7 @@ test('safeParse should parse correct data', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertSuccessfulResult(result, content, assertOpts);
 });
@@ -28,7 +28,7 @@ test('safeParse should reject string too small', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
@@ -50,7 +50,7 @@ test('safeParse should reject string too big', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
@@ -72,7 +72,7 @@ test('safeParse should reject incorrect type', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
@@ -94,7 +94,7 @@ test('safeParse should reject invalid enum', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
@@ -116,7 +116,7 @@ test('safeParse should reject invalid url', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
@@ -142,7 +142,7 @@ test('safeParse should reject invalid literal', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
@@ -163,7 +163,7 @@ test('safeParse should reject number under the minimum', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
@@ -190,7 +190,7 @@ test('safeParse should reject number that are not finite', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
@@ -224,7 +224,7 @@ test('safeParse should reject invalid discriminatedUnion', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
@@ -248,7 +248,7 @@ test('safeParse should reject invalid union', () => {
   };
   const result = safeParse<TestSchema>(content, {
     schema,
-    formatting: 'standard',
+    formatting: 'human-friendly',
   });
   assertFailedResult(
     result,
